Rename JournalEntry component
