+++ components/JournalEntry.js
