+++ pages/work-journal.js
@@ -2,7 +2,7 @@ import Head from "next/head";
 import VisibilitySensor from "react-visibility-sensor";
 
 import { getAllJournalEntries } from "../lib/api";
-import JournalEntry from "../components/journal-entry";
+import JournalEntry from "../components/JournalEntry";
 
 export default function WorkJournal({ journalEntries }) {
   return (
@@ -11,14 +11,14 @@ export default function WorkJournal({ journalEntries }) {
         <title>Work Journal - Darin Doria</title>
         <meta
           name="description"
-          content="A weekly record of notable things going on at work, with side projects or general learnings."
+          content="A weekly record of notable things going on with work and side projects, plus general learnings."
         />
         <link
           rel="stylesheet"
           href="https://unpkg.com/swiper/swiper-bundle.min.css"
         />
       </Head>
-      <div className="mt-8">
+      <div className="pt-8">
         <h3 className="mt-2 mx-auto text-3xl leading-8 font-extrabold tracking-tight text-gray-900 w-full text-center md:max-w-screen-sm lg:max-w-screen-md xl:max-w-screen-lg sm:text-4xl sm:leading-10 dark:text-gray-200">
           Work Journal
         </h3>
